requirements.txt
pandas
kafka-python
psycopg2-binary
